Fix video processing logic and type conversion in media handling
Updated the type conversion for ffmpeg command to prevent runtime issues and corrected the order of video preview and thumbnail generation. Added a call to save media updates in the database for improved data consistency.

+++ video.go
@@ -147,7 +147,7 @@ func GenerateVideoThumbnail(media *Media) error {
 
 	cmd := exec.Command("ffmpeg",
 		"-y",
-		"-ss", fmt.Sprintf("%.2f", midpoint),
+		"-ss", fmt.Sprintf("%.2f", float64(midpoint)),
 		"-i", absInput,
 		"-vframes", "1",
 		"-q:v", "2", // High quality JPEG
